Separated the classes
