+++ Data-recording-UI/Controller.py
@@ -1,5 +1,10 @@
 import seatease.spectrometers as s  # Emulator to test w/o spectrometer
 
+from LoginPage import *
+from SignUp import *
+from ResetPW import *
+from LogPatient import *
+
 import tkinter as tk
 from datetime import datetime
 from Design import *
@@ -31,7 +36,6 @@ ymax = np.around(max(data), decimals=2)
 # minIntTime =spec.minimum_integration_time_micros          UNCOMMENT
 
 
-
 class Controller(tk.Tk):
     def __init__(self, ax, *args, **kwargs):
         tk.Tk.__init__(self, *args, **kwargs)
@@ -51,298 +55,47 @@ class Controller(tk.Tk):
         # Dictionary of class names, frames (Ex: HomePage, actual frame)
         self.frames = {}
 
-        # Adds pages to the container
-        for frame in (LoginPage, LogPatient, DataRecording, SignUp, ResetPW):
-            if frame is DataRecording:
-                print("I am DataRecording")
-                tempFrame = frame(container, self, ax)
-                self.frames[frame] = tempFrame
+        self.frames["LoginPage"] = LoginPage(container, self)
+        self.frames["LoginPage"].grid(row=0, column=0, sticky="nsew")
 
-                tempFrame.grid(row=0, column=0, sticky="nsew")
-            # Initialize frames and save them into frames
-            else:
-                print("I am the rest frames")
-                tempFrame = frame(container, self)
-                self.frames[frame] = tempFrame
+        self.frames["LogPatient"] = LogPatient(container, self)
+        self.frames["LogPatient"].grid(row=0, column=0, sticky="nsew")
 
-                tempFrame.grid(row=0, column=0, sticky="nsew")
+        self.frames["DataRecording"] = DataRecording(container, self, ax)
+        self.frames["DataRecording"].grid(row=0, column=0, sticky="nsew")
+
+        self.frames["SignUp"] = SignUp(container, self)
+        self.frames["SignUp"].grid(row=0, column=0, sticky="nsew")
+
+        self.frames["ResetPW"] = ResetPW(container, self)
+        self.frames["ResetPW"].grid(row=0, column=0, sticky="nsew")
+
 
         # Starting page
-        self.show_frame(LoginPage)
+        self.show_login_frame()
         # endregion
 
-    def get_DataRecording(self):
-        return self.all_frames[DataRecording]   # DataRecording frame
 
     # When called, passes the frame or page to be showed on windows
-    def show_frame(self, cont):
-        frame = self.frames[cont]
+    def show_login_frame(self):
+        frame = self.frames["LoginPage"]
         frame.tkraise()
 
+    def show_patientLog_frame(self):
+        frame = self.frames["LogPatient"]
+        frame.tkraise()
 
-class LoginPage(tk.Frame):
+    def show_dataRecording_frame(self):
+        frame = self.frames["DataRecording"]
+        frame.tkraise()
 
-    def __init__(self, parent, controller):
-        tk.Frame.__init__(self, parent)
+    def show_signUp_frame(self):
+        frame = self.frames["SignUp"]
+        frame.tkraise()
 
-        password_bullets = "\u2022"  # Bullet points for password security
-
-        # region Design
-        welcome_label = ttk.Label(self, text="Welcome to the BMI reading Platform", font=LARGE_FONT)
-        welcome_label.pack(pady=10, padx=10)
-
-        # Picture for paths up
-        paws_up_image = ImageTk.PhotoImage(Image.open("./images/PathsUp.gif"), master=self)
-        paws_up_image_label = tk.Label(self, image=paws_up_image)
-        paws_up_image_label.image = paws_up_image
-        paws_up_image_label.pack()
-
-        # Email Label and entry box
-        email_label = ttk.Label(self, text="Email *", font=SMALL_FONT)
-        email_label.pack(pady=10, padx=10)
-        email_entry = ttk.Entry(self)
-        email_entry.insert(0, "test@yahoo.com")  # 2TEST ERASE!!!
-        email_entry.pack()
-
-        # Password label and entry box
-        password_label = ttk.Label(self, text="Password *", font=SMALL_FONT)
-        password_label.pack(pady=10, padx=10)
-        password_entry = ttk.Entry(self, show=password_bullets)
-        password_entry.insert(0, "badbunny")  # 2TEST ERASE!!!
-        password_entry.pack()
-
-        # Login Button
-        log_in_button = ttk.Button(self, text="Log In", command=lambda: check_credentials())
-        log_in_button.pack(pady=10)
-
-        # Signup Button
-        sign_up_button = ttk.Button(self, text="Sign Up", command=lambda: controller.show_frame(SignUp))
-        sign_up_button.pack()
-
-        # Forgot Password Button
-        forgot_password_button = ttk.Button(self, text="Forgot Password",
-                                            command=lambda: controller.show_frame(ResetPW))
-        forgot_password_button.pack(pady=20, padx=10)
-
-        # endregion
-
-        def check_credentials():
-            if hm.check_fields_inputs(
-                    emailEntry=email_entry,
-                    passwordEntry=password_entry,
-                    checkEmailFormat=email_entry.get()):
-                controller.show_frame(LogPatient)
-
-
-class SignUp(tk.Frame):
-    first_name = ""
-    middle_Initial = ""
-    last_name = ""
-    email = ""
-    institution = ""
-
-    def __init__(self, parent, controller):
-        tk.Frame.__init__(self, parent)
-
-        # region Design
-        welcome_label = ttk.Label(self, text="New Researcher", font=LARGE_FONT)
-        welcome_label.pack(pady=10)
-
-        f_name_label = ttk.Label(self, text="First Name *", font=SMALL_FONT)
-        f_name_label.pack(pady=10)
-        f_name_entry = ttk.Entry(self)
-        f_name_entry.pack()
-
-        m_initial_label = ttk.Label(self, text="Middle Initial *", font=SMALL_FONT)
-        m_initial_label.pack(pady=10)
-        middle_initial_entry = ttk.Entry(self)
-        middle_initial_entry.pack()
-
-        l_name_label = ttk.Label(self, text="Last Name *", font=SMALL_FONT)
-        l_name_label.pack(pady=10)
-        l_name_entry = ttk.Entry(self)
-        l_name_entry.pack()
-
-        email_label = ttk.Label(self, text="Email *", font=SMALL_FONT)
-        email_label.pack(pady=10)
-        email_entry = ttk.Entry(self)
-        email_entry.pack()
-
-        inst_label = ttk.Label(self, text="Institution *", font=SMALL_FONT)
-        inst_label.pack(pady=10)
-        inst_entry = ttk.Entry(self)
-        inst_entry.pack()
-
-        password_label = ttk.Label(self, text="Password *", font=SMALL_FONT)
-        password_label.pack(pady=10)
-        password_entry = ttk.Entry(self, show="*")
-        password_entry.pack()
-
-        retype_password_label = ttk.Label(self, text="Retype password *", font=SMALL_FONT)
-        retype_password_label.pack(pady=10)
-        retype_password_entry = ttk.Entry(self, show="*")
-        retype_password_entry.pack()
-
-        sign_up_button = ttk.Button(self, text="Sign Up", command=lambda: signUp())
-        sign_up_button.pack(pady=10)
-
-        alreadyHaveAnAccount = ttk.Button(self, text="Already have an Account? Login",
-                                          command=lambda: controller.show_frame(LoginPage))
-        alreadyHaveAnAccount.pack()
-
-        # endregion
-
-        # region Methods
-        def signUp():
-            # check if fields are empty, if password match and if email is in correct format
-            if hm.check_fields_inputs(
-                    fNameEntry=f_name_entry,
-                    middleInitialEntry=middle_initial_entry,
-                    lNameEntry=l_name_entry,
-                    instEntry=inst_entry,
-                    passwordEntry=password_entry,
-                    reEnterPasswordEntry=retype_password_entry,
-                    emailEntry=email_entry,
-                    checkEmailFormat=email_entry.get(),
-            ):
-                get_values()
-                print(self.first_name)
-                print(self.middle_Initial)
-                print(self.last_name)
-                print(self.institution)
-
-        def get_values():
-            self.first_name = f_name_entry.get()
-            self.middle_Initial = middle_initial_entry.get()
-            self.last_name = l_name_entry.get()
-            self.email = email_entry.get()
-            self.institution = inst_entry.get()
-        # endregion
-
-
-class ResetPW(tk.Frame):
-
-    def __init__(self, parent, controller):
-        tk.Frame.__init__(self, parent)
-
-        # region Design
-        welcome_label = ttk.Label(self, text="Forgot Your Password?", font=LARGE_FONT)
-        welcome_label.pack(pady=20, padx=10)
-
-        f_name_label = ttk.Label(self, text="Enter registered email", font=SMALL_FONT)
-        f_name_label.pack(pady=10, padx=10)
-        f_name_entry = ttk.Entry(self)
-        f_name_entry.pack()
-
-        send_button = ttk.Button(self, text="Send password reset", command=lambda: controller.show_frame(LoginPage))
-        send_button.pack(pady=30, padx=10)
-        # endregion
-
-
-class LogPatient(tk.Frame):
-    # values for all entries
-    age_value = 0
-    weight_value = 0
-    height_value = 0
-    skin_color_type = ""
-    ethnicity_option_selected = ""
-    gender_option_selected = ""
-    duration_value = 0
-
-    def __init__(self, parent, controller):
-        tk.Frame.__init__(self, parent)
-
-        # region Design
-        label = ttk.Label(self, text="Patient's Info", font=LARGE_FONT)
-        label.grid(row=0, column=1, padx=0, pady=10)
-
-        id_label = ttk.Label(self, text="Patient ID:", font=SMALL_FONT)
-        id_label.grid(row=2, column=0, padx=0, pady=10)
-        id_val = ttk.Label(self, text="001", font=SMALL_FONT)  # Interactive get subjID from DB
-        id_val.grid(row=2, column=1)
-
-        age_label = ttk.Label(self, text="Age:", font=SMALL_FONT)
-        age_label.grid(row=4, column=0, padx=10, pady=10)
-        age_entry = ttk.Entry(self)
-        age_entry.insert(0, 30)  # 2TEST ERASE!!!
-        age_entry.grid(row=4, column=1)
-
-        gender_label = ttk.Label(self, text="Sex:", font=SMALL_FONT)
-        gender_label.grid(row=6, column=0, padx=10, pady=10)
-        self.gender_option_selected = tk.StringVar()
-        self.gender_option_selected.set(hm.Gender[0])  # Initial value
-        gender_options = ttk.OptionMenu(self, self.gender_option_selected, *hm.Gender)
-        gender_options.grid(row=6, column=1)
-        print(f'This is gender first value: {self.gender_option_selected}')
-
-        weight_label = ttk.Label(self, text="Weight:", font=SMALL_FONT)
-        weight_label.grid(row=8, column=0, padx=10, pady=10)
-        weight_entry = ttk.Entry(self)
-        weight_entry.insert(0, 130)  # 2TEST ERASE!!!
-        weight_entry.grid(row=8, column=1)
-        weight_label_unit = ttk.Label(self, text="Lb", font=SMALL_FONT)
-        weight_label_unit.grid(row=8, column=2, padx=10, pady=10)
-
-        height_label = ttk.Label(self, text="Height:", font=SMALL_FONT)
-        height_label.grid(row=10, column=0, padx=10, pady=10)
-        height_entry = ttk.Entry(self)
-        height_entry.insert(0, 5.1)  # 2TEST ERASE!!!
-        height_entry.grid(row=10, column=1)
-
-        ethnicity_label = ttk.Label(self, text="Ethnicity:", font=SMALL_FONT)
-        ethnicity_label.grid(row=12, column=0, padx=10, pady=10)
-        self.ethnicity_option_selected = tk.StringVar()
-        self.ethnicity_option_selected.set(hm.Ethnicity[0])  # Initial value
-        ethnicity_options = ttk.OptionMenu(self, self.ethnicity_option_selected, *hm.Ethnicity)
-        ethnicity_options.grid(row=12, column=1)
-
-        skin_color_label = ttk.Label(self, text="Fitzpatrick Scale:", font=SMALL_FONT)
-        skin_color_label.grid(row=16, column=0, padx=10, pady=10)
-        self.skin_color_type = tk.StringVar()
-        self.skin_color_type.set(hm.SkinColor[0])  # Initial value
-        skin_color_entry = ttk.OptionMenu(self, self.skin_color_type, *hm.SkinColor)
-        skin_color_entry.grid(row=16, column=1)
-
-        save_button = ttk.Button(self, text="Save and Continue", command=lambda: save_and_go_to_recording_page())
-        save_button.grid(row=26, column=1, padx=10, pady=30)
-
-        # endregion
-
-        # region methods
-        def save_and_go_to_recording_page():
-
-            if get_values():
-                print(self.age_value)
-                print(self.height_value)
-                print(self.weight_value)
-                print(self.ethnicity_option_selected.get())
-                print(self.gender_option_selected.get())
-                print(self.skin_color_type.get())
-                # move to recording page
-                controller.show_frame(DataRecording)
-
-            return
-
-        # get_values stores all values from fields into variables and returns any errors found when trying to
-        # convert each field into its respective type
-        def get_values():
-
-            if hm.check_fields_inputs(
-                    ageEntry=age_entry,
-                    heightEntry=height_entry,
-                    weightEntry=weight_entry,
-                    ethnicityOption=self.ethnicity_option_selected.get(),
-                    genderOption=self.gender_option_selected.get(),
-                    skinColorOption=self.skin_color_type.get()):
-
-                self.age_value = int(age_entry.get())
-                self.height_value = float(height_entry.get())
-                self.weight_value = int(weight_entry.get())
-                return True
-            else:
-                return False
-        # endregion
-
+    def show_resetPW_frame(self):
+        frame = self.frames["SignUp"]
+        frame.tkraise()
 
 class DataRecording(tk.Frame):
     body_part_option_selected = ""
@@ -358,7 +111,7 @@ class DataRecording(tk.Frame):
     incident = np.ones(len(x))  # dummy values to prevent error in Absorbance when no dark recorded
     AbMode = 0  # start in raw intensity mode
 
-    def __init__(self, parent, controller, ax=None):
+    def __init__(self, parent, controller, ax):
         global data, x
         global IntTime, Averages
         global xmin, xmax, ymin, ymax
@@ -452,10 +205,10 @@ class DataRecording(tk.Frame):
         btn_pause_resume.pack(side='top', pady=10)
 
         diff_patient_button = ttk.Button(self.frame2, text="Next Patient",
-                                         command=lambda: controller.show_frame(LogPatient))
+                                         command=lambda: controller.show_frame("LogPatient"))
         diff_patient_button.pack(side='top', pady=10)
 
-        log_out_button = tk.Button(self.frame1, text="Log out", command=lambda: controller.show_frame(LoginPage))
+        log_out_button = tk.Button(self.frame1, text="Log out", command=lambda: controller.show_login_frame())
         log_out_button.pack(side='bottom', pady=10)
 
         checkbox_value = tk.IntVar()
@@ -746,5 +499,5 @@ class DataRecording(tk.Frame):
 
 fig, ax = plt.subplots()
 app = Controller(ax)
-ani = animation.FuncAnimation(fig, app.frames[DataRecording].update, interval=10, blit=False)
+ani = animation.FuncAnimation(fig, app.frames["DataRecording"].update, interval=10, blit=False)
 app.mainloop()
